voodoo: Use 32bit message serials.
Sorry for breaking compatibility...

+++ lib/voodoo/types.h
@@ -34,7 +34,7 @@
 
 typedef u32 VoodooInstanceID;
 typedef u32 VoodooMethodID;
-typedef u64 VoodooMessageSerial;
+typedef u32 VoodooMessageSerial;
 
 #define VOODOO_INSTANCE_NONE  ((VoodooInstanceID) 0)
